Correct reference for VGR
We judged the Fihn method to be better than the Cannegieter method,
because it divides by the number of intervals instead of
by the number of INRs.
The difference will be marginal on large samples; on smaller samples
we think Fihn is more appropriate, especially when the periods are
not of equal size.

Turns out I was already (by accident) using Fihn instead of Cannegieter.
This change was "introduced" in the JasperTools-package; previous
versions of the VGR calculations (including qoac_statistics) used the
Cannegieter method.